GetPlayerKeys up/down & left/right keys
Casted upDown & leftRight keys to signed.

+++ Server/Components/Pawn/Scripting/Player/Natives.cpp
@@ -682,9 +682,11 @@ SCRIPT_API(GetPlayerCameraMode, int(IPlayer& player))
 SCRIPT_API(GetPlayerKeys, bool(IPlayer& player, int& keys, int& updown, int& leftright))
 {
     const PlayerKeyData& keyData = player.getKeyData();
+
     keys = keyData.keys;
-    updown = keyData.upDown;
-    leftright = keyData.leftRight;
+    updown = static_cast<int16_t>(keyData.upDown);
+    leftright = static_cast<int16_t>(keyData.leftRight);
+
     return true;
 }
